Move load tasks to module.
